requirements.txt
Flask>=2.2
python-dotenv>=0.20
google-genai>=0.1.0
